First project launch !
For now, I've just created a file with a first project: that of recreating a 4-operation calculator esthetically with python!
I've placed a “journal.txt” file where I explain what I'm doing and planning to do, how and with what I'm programming this project.
Currently, the code only contains what to display “A” or “B” with 2 different buttons. This will enable me, later on, to display numbers, operations and calculation results!